refactor: remove extra layer of builder
